backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
mysqlclient==2.2.4
dj-database-url==2.2.0
gunicorn==22.0.0
